glglglgl

+++ server/routes-v2/user/repos/utils.js
@@ -2,6 +2,7 @@ const conversion_queue = require("../../../file-conversion");
 const {logger} = require("../../../logger");
 const fs = require("fs");
 const {as_data_string} = require("../../../db_utils");
+const {Item} = require("../../../database/item");
 
 const upload_in_progress = {};
 
@@ -30,11 +31,15 @@ async function finalize_file_upload(file_path, metadata, repos, user, file_hash)
             break;
     }
 
-    const parent_directory = (await Directories.find_or_create(repos.id, meta.virtual_path, {owner: user.id}));
+    const parent_directory = (await Item.find_or_create_directory_from_path(repos.id, meta.virtual_path, {owner: user.id}));
 
     // Ensure the file doesn't already exists
-    const existing_file = await File.from_path(repos.id, meta.virtual_path + "/" + meta.file_name);
+    const existing_file = await Item.from_path(repos.id, meta.virtual_path + "/" + meta.file_name);
     if (existing_file) {
+        if (!existing_file.is_regular_file) {
+            logger.error("There already is a folder at this path")
+            return null;
+        }
         logger.warn(`File ${JSON.stringify(metadata)} with the same name already exists, but with different data inside. Replacing with new one`);
         fs.renameSync(file_path, existing_file.storage_path());
         existing_file.size = meta.file_size;
@@ -46,9 +51,10 @@ async function finalize_file_upload(file_path, metadata, repos, user, file_hash)
     }
 
     try {
-        const file_meta = await new File({
+        const file_meta = await new Item({
             repos: repos.id,
             owner: user.id,
+            is_regular_file: true,
             name: as_data_string(meta.file_name),
             description: as_data_string(meta.description),
             mimetype: meta.mimetype,
@@ -59,10 +65,11 @@ async function finalize_file_upload(file_path, metadata, repos, user, file_hash)
         }).push();
 
         fs.renameSync(file_path, file_meta.storage_path());
+        console.log("B");
         return file_meta;
     }
     catch (error) {
-        logger.error('file insertion failed : ' + error);
+        logger.error('file insertion failed : ', error);
         return null;
     }
 }
